add missing test files
